requirements.txt
langchain==0.3.26
flask==3.1.1
pypdf==5.6.1
sentence-transformers==4.1.0
python_dotenv==1.1.0
langchain-pinecone==0.2.8
langchain_groq
langchain_community==0.3.26
ipykernel
langchain_huggingface
pinecone

-e.
